requirements.txt
tequila-basic
numpy

autograd
openfermion ~= 1.0 # can not be smaller than 1.0
qulacs

setuptools
